% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeopardy_clues.R
\docType{data}
\name{jeopardy_clues}
\alias{jeopardy_clues}
\title{Jeopardy! clues}
\format{
A data frame with 98,333 rows and 7 variables:
\itemize{
\item \code{date}: The episode's airdate.
\item \code{category}: The clue's category.
\item \code{comment}: Any helpful comments announced to explain the category.
\item \code{value}: The clue's dollar value. If \code{NA}, the clue is either a Daily
Double or Final Jeopardy question, where contestants can wager a dollar
value, or a Tiebreaker question, which has no dollar value.
\item \code{clue}: The clue text or "answer".
\item \code{response}: The correct response, which should be phrased in the form of
a question.
\item \code{link}: A link to a visual or audio file which appeared alongside the clue.
}
}
\source{
http://www.j-archive.com/
}
\usage{
jeopardy_clues
}
\description{
A dataset containing clues and correct responses for 2660 games of
\emph{Jeopardy!} since 2013.
}
\keyword{datasets}
